+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # v0.11.5 (Upcoming Release)
 
 - Initial implementation of the robust hat matrix regression estimator
+- Add more test to robust hat matrix regression estimator
 - Introduce `view` in LTS and LMS.
+
 
 # v0.11.4